backend/requirements.txt
langchain
ddgs
langgraph
langchain-community
langchain-core
langchain-openai
fastapi
pydantic
uvicorn[standard]
aiosqlite
markdown
fpdf2
uagents
uagents-core
python-dotenv